Move meta into the operations module
