+++ memorious/operations/meta.py
